# packages/detensor-cli/detensor_cli-0.1.9.tar.gz/detensor_cli-0.1.9/detensor_cli/template/pro_project/storage-tt/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "storage API是储存层的对象储存API",
        "title": "storage API",
        "contact": {
            "name": "bravomikekilo",
            "email": "bravomikekilo@buaa.edu.cn"
        },
        "version": "0.1"
    },
    "basePath": "/api",
    "paths": {
...
        "/core/v1/namespace/{namespace}/{kind}": {
            "get": {
                "description": "获取命名空间中某个类型的全部对象的接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Object 对象"
                ],
                "summary": "获取命名空间中某个类型的全部对象的接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "命名空间的名称",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "对象的类型名称",
                        "name": "kind",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    }
                }
            }
        },
        "/core/v1/namespace/{namespace}/{kind}/{name}": {
            "get": {
                "description": "根据命名空间和对象的类型或名称获取对象",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Object 对象"
                ],
                "summary": "根据命名空间和对象的类型与名称获取对象",
                "parameters": [
                    {
                        "type": "string",
                        "description": "命名空间的名称",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "对象的类型",
                        "name": "kind",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "对象的类型",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "对象的版本",
                        "name": "version",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "如果对象不使用版本, 返回JSON编码的对象, 否则返回版本为键,对象为值的字典",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "找不到对应类型,名称的对象"
                    }
                }
            }
        },
...
        "/core/v1/object/{uid}": {
            "get": {
                "description": "获取命名空间中某个类型的全部对象的接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Object 对象"
                ],
                "summary": "根据UID查询单个对象",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "对象的UID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JSON编码的对象",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "找不到拥有该UID的对象"
                    }
                }
            }
        },
...
        "/core/v1/{kind}": {
            "get": {
                "description": "获取某个类型的全部对象的接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Object 对象"
                ],
                "summary": "获取某个类型的全部对象的接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "对象的类型名称",
                        "name": "kind",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    }
                }
            }
        },
        "/core/v1/{kind}/{name}": {
            "get": {
                "description": "根据对象的类型或名称获取全局对象",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Object 对象"
                ],
                "summary": "根据对象的类型与名称获取全局对象",
                "parameters": [
                    {
                        "type": "string",
                        "description": "对象的类型",
                        "name": "kind",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "对象的名称或uid",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "对象的版本",
                        "name": "version",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "如果对象不使用版本, 返回JSON编码的对象, 否则返回版本为键,对象为值的字典",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "找不到对应类型,名称的对象"
                    }
                }
            }
        }
    },
...
}